fix: mock data treeshake
