+++ apps/web/src/entities/exercise/api/index.ts
@@ -1,4 +1,3 @@
-import { FitnessMockData } from '@entities/fitness/api'
 import useCreateExerciseByExercisePreset from './useCreateExerciseByExercisePreset'
 import useCreateExerciseBySchedule from './useCreateExerciseBySchedule'
 import useDeleteExerciseById from './useDeleteExerciseById'
@@ -20,16 +19,3 @@ export {
   useUpdateExerciseListByScheduleId,
   useUpdateExerciseListByExercisePresetId,
 }
-
-export const ExerciseMockData: { [key: number]: ExerciseDataStoreType } = Array(20).fill(0).reduce((acc, _cur, i) => {
-  const id = i + 1
-  const temp: ExerciseDataStoreType = {
-    deps: 0,
-    exercise: id,
-    id,
-    fitness: FitnessMockData[i],
-    __typename: 'ExerciseWithFitness'
-  }
-  acc[id] = temp
-  return acc
-}, {})
