+++ apps/web/src/entities/exercise/api/useCreateExerciseByExercisePreset.mock.ts
@@ -1,7 +1,7 @@
 import type { MockedResponse } from '@apollo/client/testing';
-import { ExerciseMockData } from '@entities/exercise/api';
+import { ExerciseMockData } from '@entities/exercise/api/exercise.mockData';
 import CreateExerciseByExercisePreset from '@entities/exercise/api/graphql/mutation/CreateExerciseByExercisePreset';
-import { FitnessMockData } from '@entities/fitness/api';
+import { FitnessMockData } from '@entities/fitness/api/fitness.mockData';
 
 const CreateExerciseByExercisePresetMock: MockedResponse<
   CreateExerciseByExercisePresetResponse,
