+++ apps/web/src/entities/exercise/api/useGetExerciseListByScheduleId.mock.ts
@@ -1,5 +1,5 @@
 import type { MockedResponse } from '@apollo/client/testing';
-import { FitnessMockData } from '@entities/fitness/api';
+import { FitnessMockData } from '@entities/fitness/api/fitness.mockData';
 import GetExerciseListByScheduleId from '../api/graphql/query/GetExerciseListByScheduleId';
 
 const GetExerciseListByScheduleIdMock: MockedResponse<
