+++ apps/web/src/entities/fitness/api/fitness.mockData.ts
@@ -1,0 +1,12 @@
+import FitnessData from '@service/Fitness/FitnessData.json'
+export const FitnessMockData: FitnessStoreType[] = FitnessData.map((v, i) => {
+  const fitness = v as FitnessStoreType
+  return {
+    ...fitness,
+    aliases: fitness.aliases || [],
+    description: fitness.description || '',
+    tips: fitness.tips || [],
+    id: i + 1,
+    __typename: 'Fitness'
+  }
+})
