+++ apps/web/src/entities/fitness/api/index.ts
@@ -1,4 +1,3 @@
-import FitnessData from '@service/Fitness/FitnessData.json'
 import useFitnessFragment from './useFitnessFragment'
 import useFitnessSimpleFragment from './useFitnessSimpleFragment'
 import useGetFitnessById from './useGetFitnessById'
@@ -16,15 +15,3 @@ export {
   useGetFitnessSimpleById,
   useGetFitnessSimpleListByIds
 }
-
-export const FitnessMockData: FitnessStoreType[] = FitnessData.map((v, i) => {
-  const fitness = v as FitnessStoreType
-  return {
-    ...fitness,
-    aliases: fitness.aliases || [],
-    description: fitness.description || '',
-    tips: fitness.tips || [],
-    id: i + 1,
-    __typename: 'Fitness'
-  }
-})
