+++ apps/web/src/entities/fitness/api/useGetFitnessSimpleListByIds.mock.ts
@@ -1,5 +1,5 @@
 import type { MockedResponse } from '@apollo/client/testing';
-import { FitnessMockData } from '@entities/fitness/api';
+import { FitnessMockData } from '@entities/fitness/api/fitness.mockData';
 import GetFitnessSimpleByIds from '@entities/fitness/api/graphql/query/GetFitnessSimpleByIds';
 
 const useGetFitnessSimpleListByIdsMock: MockedResponse<
